+++ src/sc_neurocore/memristor/__init__.py
@@ -11,4 +11,56 @@
 Tier: research.
 """
 
+from .memristor_mapper import (
+    AgingReport,
+    AgingSimulator,
+    CompensationLUT,
+    CompensationStrategy,
+    ConductanceModel,
+    CrossbarArray,
+    CrossbarEstimator,
+    CrossbarMapping,
+    CrossbarPowerEstimate,
+    CrossbarTopology,
+    IRDropModel,
+    MappingResult,
+    MemristorMapper,
+    MemristorTechnology,
+    MonteCarloReport,
+    MonteCarloSimulator,
+    SCAbsorbEncoder,
+    SneakPathModel,
+    StuckFaultMap,
+    VariabilityInjector,
+    VerilogEmitter,
+    WriteVerifyProtocol,
+    WriteVerifyResult,
+)
+
 __tier__ = "research"
+
+__all__ = [
+    "AgingReport",
+    "AgingSimulator",
+    "CompensationLUT",
+    "CompensationStrategy",
+    "ConductanceModel",
+    "CrossbarArray",
+    "CrossbarEstimator",
+    "CrossbarMapping",
+    "CrossbarPowerEstimate",
+    "CrossbarTopology",
+    "IRDropModel",
+    "MappingResult",
+    "MemristorMapper",
+    "MemristorTechnology",
+    "MonteCarloReport",
+    "MonteCarloSimulator",
+    "SCAbsorbEncoder",
+    "SneakPathModel",
+    "StuckFaultMap",
+    "VariabilityInjector",
+    "VerilogEmitter",
+    "WriteVerifyProtocol",
+    "WriteVerifyResult",
+]